refactor: move EntityCreatorEditor into folder
