+++ src/components/modals/EntityCreatorEditor/EntityCreatorEditorPresentation.tsx
@@ -5,14 +5,14 @@
 import * as React from 'react'
 import { Modal } from 'office-ui-fabric-react/lib/Modal'
 import * as OF from 'office-ui-fabric-react'
-import * as TC from '../tipComponents'
-import ActionDetailsList from '../ActionDetailsList'
-import ConfirmCancelModal from './ConfirmCancelModal'
-import { CLDropdownOption } from './CLDropDownOption'
-import * as ToolTip from '../ToolTips'
+import * as TC from '../../tipComponents'
+import ActionDetailsList from '../../ActionDetailsList'
+import ConfirmCancelModal from '../ConfirmCancelModal'
+import { CLDropdownOption } from '../CLDropDownOption'
+import * as ToolTip from '../../ToolTips'
 import { ActionBase } from '@conversationlearner/models'
 import './EntityCreatorEditor.css'
-import { FM } from '../../react-intl-messages'
+import { FM } from '../../../react-intl-messages'
 import { InjectedIntlProps, FormattedMessage } from 'react-intl'
 
 interface Props extends InjectedIntlProps {
