+++ src/components/modals/EntityCreatorEditor/index.ts
@@ -1,0 +1,3 @@
+import Component from './EntityCreatorEditor'
+
+export default Component
